+++ src/eve-server/ship/dgmtypeattributeinfo.cpp
@@ -43,8 +43,8 @@ dgmtypeattributemgr::dgmtypeattributemgr()
     DgmTypeAttributeSet * entry = NULL;
     DBResultRow row;
 
-    int amount = res.GetRowCount();
-    for (int i = 0; i < amount; i++)
+    size_t amount = res.GetRowCount();
+    for (size_t i = 0; i < amount; i++)
     {
         res.GetRow(row);
         uint32 typeID = row.GetUInt(0);
